Revert "all the filter functions are in a separate file AND THEY WORK"
This reverts commit 52a395f1d90f2e05681656f8dd6d3ea045e5264a.
