+++ index.html
@@ -97,7 +97,6 @@
     </footer>
     <script type="module" src="./scripts/home.js"></script>
     <script type="module" src="./scripts/placesFunctions.js"></script>
-    <script type="module" src="./scripts/filteringFunctions.js"></script>
 </body>
 
 </html>
